Utilize JSROOT.makeImage now that 'blob:' is not blocked by CSP
