+++ QualityControl/public/common/utils.js
@@ -13,6 +13,21 @@
  */
 
 import { isUserRoleSufficient } from '../../../../library/userRole.enum.js';
+import { generateDrawingOptionString } from '../../library/qcObject/utils.js';
+
+/* global JSROOT */
+
+/**
+ * Map of allowed `ROOT.makeImage` file extensions to MIME types
+ * @type {Map<string, string>}
+ */
+const SUPPORTED_ROOT_IMAGE_FILE_TYPES = new Map([
+  ['svg', 'image/svg+xml'],
+  ['png', 'file/png'],
+  ['jpg', 'file/jpeg'],
+  ['jpeg', 'file/jpeg'],
+  ['webp', 'file/webp'],
+]);
 
 /**
  * Generates a new ObjectId
@@ -182,4 +197,41 @@ export const triggerDownload = (url, filename) => {
   link.href = url;
   link.download = filename;
   link.click();
+};
+
+/**
+ * Downloads a file
+ * @param {Blob|MediaSource} file - The file to download
+ * @param {string} filename - The name of the file including the file extension
+ */
+export const downloadFile = (file, filename) => {
+  const url = URL.createObjectURL(file);
+  try {
+    triggerDownload(url, filename);
+  } finally {
+    URL.revokeObjectURL(url);
+  }
+};
+
+/**
+ * Generates a rasterized image of a JSROOT RootObject and triggers download.
+ * @param {string} filename - The name of the downloaded file including its extension.
+ * @param {RootObject} root - The JSROOT RootObject to render.
+ * @param {string[]} [drawingOptions=[]] - Optional array of JSROOT drawing options.
+ */
+export const downloadRoot = async (filename, root, drawingOptions = []) => {
+  const filetype = filename.substring(filename.lastIndexOf('.') + 1).toLowerCase();
+  const mime = SUPPORTED_ROOT_IMAGE_FILE_TYPES.get(filetype);
+  if (!mime) {
+    throw new Error(`The file extension (${filetype}) is not supported`);
+  }
+
+  const image = await JSROOT.makeImage({
+    object: root,
+    option: generateDrawingOptionString(root, drawingOptions),
+    format: filetype,
+    as_buffer: true,
+  });
+  const blob = new Blob([image], { type: mime });
+  downloadFile(blob, filename);
 };
