+++ QualityControl/public/object/objectTreePage.js
@@ -20,7 +20,7 @@ import virtualTable from './virtualTable.js';
 import { defaultRowAttributes, qcObjectInfoPanel } from '../common/object/objectInfoCard.js';
 import { downloadButton } from '../common/downloadButton.js';
 import { resizableDivider } from '../common/resizableDivider.js';
-import { downloadRootObjectAsImage } from '../common/rootDownload.js';
+import { downloadRoot } from '../common/utils.js';
 
 /**
  * Shows a page to explore though a tree of objects with a preview on the right if clicked
@@ -95,18 +95,19 @@ function objectPanel(model) {
  */
 const drawPlot = (model, object) => {
   const { name, qcObject, validFrom, id, drawingOptions = [], displayHints = [] } = object;
+  const { root } = qcObject;
   const href = validFrom ?
     `?page=objectView&objectName=${name}&ts=${validFrom}&id=${id}`
     : `?page=objectView&objectName=${name}`;
   return h('', { style: 'height:100%; display: flex; flex-direction: column' }, [
     h('.item-action-row.flex-row.g1.p1', [
-      h('button.btn', {
+      root.fArray?.length && h('button.btn', {
         title: 'Download as PNG',
         id: 'download-image-button',
         onclick: async (event) => {
           try {
             event.target.disabled = true;
-            await downloadRootObjectAsImage(`${name}.png`, qcObject.root, [...drawingOptions, ...displayHints]);
+            await downloadRoot(`${name}.webp`, root, [...drawingOptions, ...displayHints]);
           } finally {
             event.target.disabled = false;
           }
